# openapi.yaml
openapi: 3.0.0
info:
  title: Parliamentary Bills API
  version: '1.0'
servers:
  - url: https://ibm-hackathon-api-6497-4acd6860-11ip61w7.onporter.run
paths:
  /api/bills:
    get:
      summary: Get all bills
      parameters:
        - name: house
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: string
        - name: title
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bill'
  /api/bills/{bill_id}:
    get:
      summary: Get bill by ID
      parameters:
        - name: bill_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bill details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'
  /api/bills/search:
    post:
      summary: Search bills
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bill'
    /api/bills/compare:
    post:
      summary: Compare two bills
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  properties:
                    bill_id_1:
                      type: string
                    bill_id_2:
                      type: string
                  required: [bill_id_1, bill_id_2]
                - type: object
                  properties:
                    query1:
                      type: string
                    query2:
                      type: string
                  required: [query1, query2]
      responses:
        '200':
          description: Comparison result
          content:
            application/json:
              schema:
                type: object
                properties:
                  bill_1_title:
                    type: string
                  bill_2_title:
                    type: string
                  comparison:
                    type: object
                    properties:
                      similarities:
                        type: array
                        items:
                          type: string
                      bill_1_unique:
                        type: array
                        items:
                          type: string
                      bill_2_unique:
                        type: array
                        items:
                          type: string
                      common_themes:
                        type: array
                        items:
                          type: string

  /api/bills/search_similar:
    post:
      summary: Find similar bills
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
              required: [query]
      responses:
        '200':
          description: Similar bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bill'
components:
  schemas:
    Bill:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        date:
          type: string
        house:
          type: string
        positives:
          type: array
          items:
            $ref: '#/components/schemas/Point'
        negatives:
          type: array
          items:
            $ref: '#/components/schemas/Point'
    Point:
      type: object
      properties:
        title:
          type: string
        explanation:
          type: string